+++ server/src/main/java/com/krillsson/sysapi/core/monitoring/monitors/CpuTemperatureMonitor.kt
@@ -2,6 +2,7 @@ package com.krillsson.sysapi.core.monitoring.monitors
 
 import com.krillsson.sysapi.core.domain.monitor.MonitorConfig
 import com.krillsson.sysapi.core.domain.monitor.MonitoredValue
+import com.krillsson.sysapi.core.domain.system.SystemInfo
 import com.krillsson.sysapi.core.monitoring.MetricQueryEvent
 import com.krillsson.sysapi.core.monitoring.Monitor
 import java.util.*
@@ -19,6 +20,11 @@ class CpuTemperatureMonitor(override val id: UUID, override val config: MonitorC
 
     override fun selectValue(event: MetricQueryEvent): MonitoredValue.NumericalValue? =
         selector(event.load, null)
+
+    override fun maxValue(info: SystemInfo): MonitoredValue.NumericalValue? {
+        // have no way of knowing this
+        return MonitoredValue.NumericalValue(120)
+    }
 
     override fun isPastThreshold(value: MonitoredValue.NumericalValue): Boolean {
         return value > config.threshold