+++ server/src/main/java/com/krillsson/sysapi/core/monitoring/monitors/DriveWriteRateMonitor.kt
@@ -4,6 +4,7 @@ import com.krillsson.sysapi.core.domain.drives.DriveLoad
 import com.krillsson.sysapi.core.domain.monitor.MonitorConfig
 import com.krillsson.sysapi.core.domain.monitor.MonitoredValue
 import com.krillsson.sysapi.core.domain.monitor.toNumericalValue
+import com.krillsson.sysapi.core.domain.system.SystemInfo
 import com.krillsson.sysapi.core.monitoring.MetricQueryEvent
 import com.krillsson.sysapi.core.monitoring.Monitor
 import java.util.*
@@ -25,6 +26,11 @@ class DriveWriteRateMonitor(override val id: UUID, override val config: MonitorC
 
     override fun selectValue(event: MetricQueryEvent): MonitoredValue.NumericalValue? =
         selector(event.load, config.monitoredItemId)
+
+    override fun maxValue(info: SystemInfo): MonitoredValue.NumericalValue? {
+        // have no way of knowing this
+        return null
+    }
 
     override fun isPastThreshold(value: MonitoredValue.NumericalValue): Boolean {
         return value > config.threshold