+++ server/src/main/java/com/krillsson/sysapi/core/monitoring/monitors/ExternalIpChangedMonitor.kt
@@ -3,6 +3,7 @@ package com.krillsson.sysapi.core.monitoring.monitors
 import com.krillsson.sysapi.core.domain.monitor.MonitorConfig
 import com.krillsson.sysapi.core.domain.monitor.MonitoredValue
 import com.krillsson.sysapi.core.domain.monitor.toConditionalValue
+import com.krillsson.sysapi.core.domain.system.SystemInfo
 import com.krillsson.sysapi.core.monitoring.MetricQueryEvent
 import com.krillsson.sysapi.core.monitoring.Monitor
 import java.util.*
@@ -30,6 +31,9 @@ class ExternalIpChangedMonitor(
     override val type: Type = Type.EXTERNAL_IP_CHANGED
 
     override fun selectValue(event: MetricQueryEvent): MonitoredValue.ConditionalValue? = selector(event.load, null)
+    override fun maxValue(info: SystemInfo): MonitoredValue.ConditionalValue? {
+        return MonitoredValue.ConditionalValue(true)
+    }
 
     override fun isPastThreshold(value: MonitoredValue.ConditionalValue): Boolean {
         return !value.value