+++ server/src/main/java/com/krillsson/sysapi/core/monitoring/monitors/LoadAveragesMonitors.kt
@@ -2,6 +2,7 @@ package com.krillsson.sysapi.core.monitoring.monitors
 
 import com.krillsson.sysapi.core.domain.monitor.MonitorConfig
 import com.krillsson.sysapi.core.domain.monitor.MonitoredValue
+import com.krillsson.sysapi.core.domain.system.SystemInfo
 import com.krillsson.sysapi.core.monitoring.MetricQueryEvent
 import com.krillsson.sysapi.core.monitoring.Monitor
 import java.util.*
@@ -9,12 +10,15 @@ import java.util.*
 class LoadAverageMonitorOneMinute(
     override val id: UUID,
     override val config: MonitorConfig<MonitoredValue.FractionalValue>
-) :
-    Monitor<MonitoredValue.FractionalValue>() {
+) : Monitor<MonitoredValue.FractionalValue>() {
 
     companion object {
         val selector: FractionalValueSelector = { load, _ ->
             MonitoredValue.FractionalValue(load.cpuLoad.loadAverages.oneMinute.toFloat())
+        }
+
+        val maxValueSelector: MaxValueFractionalSelector = { info, _ ->
+            MonitoredValue.FractionalValue(info.cpuInfo.centralProcessor.logicalProcessorCount.toFloat())
         }
     }
 
@@ -22,6 +26,10 @@ class LoadAverageMonitorOneMinute(
 
     override fun selectValue(event: MetricQueryEvent): MonitoredValue.FractionalValue? =
         selector(event.load, null)
+
+    override fun maxValue(info: SystemInfo): MonitoredValue.FractionalValue? {
+        return maxValueSelector(info, null)
+    }
 
     override fun isPastThreshold(value: MonitoredValue.FractionalValue): Boolean {
         return value > config.threshold
@@ -38,12 +46,20 @@ class LoadAverageMonitorFiveMinutes(
         val selector: FractionalValueSelector = { load, _ ->
             MonitoredValue.FractionalValue(load.cpuLoad.loadAverages.fiveMinutes.toFloat())
         }
+
+        val maxValueSelector: MaxValueFractionalSelector = { info, _ ->
+            MonitoredValue.FractionalValue(info.cpuInfo.centralProcessor.logicalProcessorCount.toFloat())
+        }
     }
 
     override val type: Type = Type.LOAD_AVERAGE_FIVE_MINUTES
 
     override fun selectValue(event: MetricQueryEvent): MonitoredValue.FractionalValue? =
         selector(event.load, null)
+
+    override fun maxValue(info: SystemInfo): MonitoredValue.FractionalValue? {
+        return maxValueSelector(info, null)
+    }
 
     override fun isPastThreshold(value: MonitoredValue.FractionalValue): Boolean {
         return value > config.threshold
@@ -60,12 +76,20 @@ class LoadAverageMonitorFifteenMinutes(
         val selector: FractionalValueSelector = { load, _ ->
             MonitoredValue.FractionalValue(load.cpuLoad.loadAverages.fifteenMinutes.toFloat())
         }
+
+        val maxValueSelector: MaxValueFractionalSelector = { info, _ ->
+            MonitoredValue.FractionalValue(info.cpuInfo.centralProcessor.logicalProcessorCount.toFloat())
+        }
     }
 
     override val type: Type = Type.LOAD_AVERAGE_FIFTEEN_MINUTES
 
     override fun selectValue(event: MetricQueryEvent): MonitoredValue.FractionalValue? =
         selector(event.load, null)
+
+    override fun maxValue(info: SystemInfo): MonitoredValue.FractionalValue? {
+        return maxValueSelector(info, null)
+    }
 
     override fun isPastThreshold(value: MonitoredValue.FractionalValue): Boolean {
         return value > config.threshold