+++ server/src/main/java/com/krillsson/sysapi/core/monitoring/monitors/MemorySpaceMonitor.kt
@@ -3,6 +3,7 @@ package com.krillsson.sysapi.core.monitoring.monitors
 import com.krillsson.sysapi.core.domain.monitor.MonitorConfig
 import com.krillsson.sysapi.core.domain.monitor.MonitoredValue
 import com.krillsson.sysapi.core.domain.monitor.toNumericalValue
+import com.krillsson.sysapi.core.domain.system.SystemInfo
 import com.krillsson.sysapi.core.monitoring.MetricQueryEvent
 import com.krillsson.sysapi.core.monitoring.Monitor
 import java.util.*
@@ -13,12 +14,20 @@ class MemorySpaceMonitor(override val id: UUID, override val config: MonitorConf
         val selector: NumericalValueSelector = { load, _ ->
             load.memory.availableBytes.toNumericalValue()
         }
+
+        val maxValueSelector: MaxValueNumericalSelector = { info, _ ->
+            info.memory.totalBytes.toNumericalValue()
+        }
     }
 
     override val type: Type = Type.MEMORY_SPACE
 
     override fun selectValue(event: MetricQueryEvent): MonitoredValue.NumericalValue? =
         selector(event.load, null)
+
+    override fun maxValue(info: SystemInfo): MonitoredValue.NumericalValue? {
+        return maxValueSelector(info, null)
+    }
 
     override fun isPastThreshold(value: MonitoredValue.NumericalValue): Boolean {
         return value < config.threshold