+++ server/src/main/java/com/krillsson/sysapi/core/monitoring/monitors/NetworkUpMonitor.kt
@@ -4,6 +4,7 @@ import com.krillsson.sysapi.core.domain.monitor.MonitorConfig
 import com.krillsson.sysapi.core.domain.monitor.MonitoredValue
 import com.krillsson.sysapi.core.domain.monitor.toConditionalValue
 import com.krillsson.sysapi.core.domain.network.NetworkInterfaceLoad
+import com.krillsson.sysapi.core.domain.system.SystemInfo
 import com.krillsson.sysapi.core.monitoring.MetricQueryEvent
 import com.krillsson.sysapi.core.monitoring.Monitor
 import java.util.*
@@ -25,6 +26,8 @@ class NetworkUpMonitor(
     override fun selectValue(event: MetricQueryEvent): MonitoredValue.ConditionalValue? {
         return selector(event.load, config.monitoredItemId)
     }
+
+    override fun maxValue(info: SystemInfo): MonitoredValue.ConditionalValue? = MonitoredValue.ConditionalValue(true)
 
     override fun isPastThreshold(value: MonitoredValue.ConditionalValue): Boolean {
         return !value.value