+++ server/src/main/java/com/krillsson/sysapi/core/monitoring/monitors/NetworkUploadRateMonitor.kt
@@ -4,6 +4,7 @@ import com.krillsson.sysapi.core.domain.monitor.MonitorConfig
 import com.krillsson.sysapi.core.domain.monitor.MonitoredValue
 import com.krillsson.sysapi.core.domain.monitor.toNumericalValue
 import com.krillsson.sysapi.core.domain.network.NetworkInterfaceLoad
+import com.krillsson.sysapi.core.domain.system.SystemInfo
 import com.krillsson.sysapi.core.monitoring.MetricQueryEvent
 import com.krillsson.sysapi.core.monitoring.Monitor
 import java.util.*
@@ -19,12 +20,25 @@ class NetworkUploadRateMonitor(
                 n.name.equals(monitoredItemId, ignoreCase = true) || n.mac.equals(monitoredItemId, ignoreCase = true)
             }?.speed?.sendBytesPerSecond?.toNumericalValue()
         }
+        val maxValueSelector: MaxValueNumericalSelector = { info, monitoredItemId ->
+            val nic = info.networkInterfaces.firstOrNull { n ->
+                n.name.equals(
+                    monitoredItemId,
+                    ignoreCase = true
+                ) || n.mac.equals(monitoredItemId, ignoreCase = true)
+            }
+            nic?.let { MonitoredValue.NumericalValue(it.speedBitsPerSeconds / 8) }
+        }
     }
 
     override val type: Type = Type.NETWORK_UPLOAD_RATE
 
     override fun selectValue(event: MetricQueryEvent): MonitoredValue.NumericalValue? {
         return selector(event.load, config.monitoredItemId)
+    }
+
+    override fun maxValue(info: SystemInfo): MonitoredValue.NumericalValue? {
+        return maxValueSelector(info, config.monitoredItemId)
     }
 
     override fun isPastThreshold(value: MonitoredValue.NumericalValue): Boolean {