+++ server/src/main/java/com/krillsson/sysapi/core/monitoring/monitors/ProcessMemoryMonitor.kt
@@ -3,6 +3,7 @@ package com.krillsson.sysapi.core.monitoring.monitors
 import com.krillsson.sysapi.core.domain.monitor.MonitorConfig
 import com.krillsson.sysapi.core.domain.monitor.MonitoredValue
 import com.krillsson.sysapi.core.domain.monitor.toNumericalValue
+import com.krillsson.sysapi.core.domain.system.SystemInfo
 import com.krillsson.sysapi.core.monitoring.MetricQueryEvent
 import com.krillsson.sysapi.core.monitoring.Monitor
 import java.util.*
@@ -14,12 +15,18 @@ class ProcessMemoryMonitor(override val id: UUID, override val config: MonitorCo
             val pid = monitoredItemID?.toInt()
             load.processes.firstOrNull { it.processID == pid }?.residentSetSize?.toNumericalValue()
         }
+
+        val maxValueSelector: MaxValueNumericalSelector = { info, _ ->
+            info.memory.totalBytes.toNumericalValue()
+        }
     }
 
     override val type: Type = Type.PROCESS_MEMORY_SPACE
 
     override fun selectValue(event: MetricQueryEvent): MonitoredValue.NumericalValue? =
         selector(event.load, config.monitoredItemId)
+
+    override fun maxValue(info: SystemInfo): MonitoredValue.NumericalValue? = maxValueSelector(info, null)
 
     override fun isPastThreshold(value: MonitoredValue.NumericalValue): Boolean {
         return value > config.threshold