+++ CHANGES.md
@@ -1,4 +1,7 @@
 ### Unreleased
+- added `deletePastEventsForMonitor`, `closeOngoingEventForMonitor` to the GraphQL-API
+- added `Monitor.maxValue` to the GraphQL-API. Useful when displaying monitored value in a graph.
+- added start value to past events
 
 ### 0.19.3
 - fix NPE when querying ContainerNetworkSettings while using podman in rootless mode